+++ gherkin/java/src/main/java/io/cucumber/gherkin/GherkinDialect.java
@@ -7,13 +7,15 @@ import com.eclipsesource.json.JsonValue;
 import java.util.ArrayList;
 import java.util.List;
 
-public class GherkinDialect {
+import static java.util.Objects.requireNonNull;
+
+public final class GherkinDialect {
     private final JsonObject keywords;
-    private String language;
+    private final String language;
 
     public GherkinDialect(String language, JsonObject keywords) {
-        this.language = language;
-        this.keywords = keywords;
+        this.language = requireNonNull(language);
+        this.keywords = requireNonNull(keywords);
     }
 
     public List<String> getFeatureKeywords() {