+++ gherkin/java/src/main/java/io/cucumber/gherkin/GherkinDialectProvider.java
@@ -13,15 +13,17 @@ import java.util.List;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Collections.sort;
 import static java.util.Collections.unmodifiableList;
+import static java.util.Objects.requireNonNull;
 
-public class GherkinDialectProvider implements IGherkinDialectProvider {
-    private static JsonObject DIALECTS;
+public final class GherkinDialectProvider {
+    private static final JsonObject DIALECTS;
     private final String defaultDialectName;
 
     public static final String JSON_PATH = "/io/cucumber/gherkin/gherkin-languages.json";
 
     static {
-        try (Reader reader = new InputStreamReader(GherkinDialectProvider.class.getResourceAsStream(JSON_PATH), UTF_8)) {
+        try (Reader reader = new InputStreamReader(GherkinDialectProvider.class.getResourceAsStream(JSON_PATH),
+                UTF_8)) {
             DIALECTS = Json.parse(reader).asObject();
         } catch (IOException e) {
             throw new GherkinException("Unable to parse " + JSON_PATH, e);
@@ -29,7 +31,7 @@ public class GherkinDialectProvider implements IGherkinDialectProvider {
     }
 
     public GherkinDialectProvider(String defaultDialectName) {
-        this.defaultDialectName = defaultDialectName;
+        this.defaultDialectName = requireNonNull(defaultDialectName);
     }
 
     public GherkinDialectProvider() {
@@ -40,8 +42,9 @@ public class GherkinDialectProvider implements IGherkinDialectProvider {
         return getDialect(defaultDialectName, null);
     }
 
-    @Override
     public GherkinDialect getDialect(String language, Location location) {
+        requireNonNull(language);
+
         JsonValue languageObject = DIALECTS.get(language);
         if (languageObject == null) {
             throw new ParserException.NoSuchLanguageException(language, location);
@@ -50,7 +53,6 @@ public class GherkinDialectProvider implements IGherkinDialectProvider {
         return new GherkinDialect(language, languageObject.asObject());
     }
 
-    @Override
     public List<String> getLanguages() {
         List<String> languages = new ArrayList<>(DIALECTS.names());
         sort(languages);