+++ html-formatter/java/src/main/java/io/cucumber/htmlformatter/MessagesToHtmlWriter.java
@@ -1,6 +1,5 @@
 package io.cucumber.htmlformatter;
 
-import io.cucumber.messages.JSON;
 import io.cucumber.messages.types.Envelope;
 
 import java.io.BufferedReader;
@@ -9,8 +8,10 @@ import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
+import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.io.Writer;
+import java.nio.charset.StandardCharsets;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Objects.requireNonNull;
@@ -19,17 +20,26 @@ import static java.util.Objects.requireNonNull;
  * Writes the message output of a test run as single page html report.
  */
 public final class MessagesToHtmlWriter implements AutoCloseable {
-
     private final String template;
-
     private final Writer writer;
+    private final Serializer serializer;
     private boolean preMessageWritten = false;
     private boolean postMessageWritten = false;
     private boolean firstMessageWritten = false;
     private boolean streamClosed = false;
 
-    public MessagesToHtmlWriter(Writer writer) throws IOException {
+    public MessagesToHtmlWriter(OutputStream outputStream, Serializer serializer) throws IOException {
+        this(
+                new OutputStreamWriter(
+                        requireNonNull(outputStream),
+                        StandardCharsets.UTF_8),
+                requireNonNull(serializer)
+        );
+    }
+
+    private MessagesToHtmlWriter(Writer writer, Serializer serializer) throws IOException {
         this.writer = writer;
+        this.serializer = serializer;
         this.template = readResource("index.mustache.html");
     }
 
@@ -67,7 +77,7 @@ public final class MessagesToHtmlWriter implements AutoCloseable {
             writer.write(",");
         }
 
-        JSON.writeValue(writer, envelope);
+        serializer.writeValue(writer, envelope);
     }
 
     /**
@@ -79,7 +89,7 @@ public final class MessagesToHtmlWriter implements AutoCloseable {
      */
     @Override
     public void close() throws IOException {
-        if(streamClosed){
+        if (streamClosed) {
             return;
         }
 
@@ -93,11 +103,15 @@ public final class MessagesToHtmlWriter implements AutoCloseable {
             writePostMessage();
             postMessageWritten = true;
         }
-        writer.close();
-        streamClosed = true;
+        try {
+            writer.close();
+        } finally {
+            streamClosed = true;
+        }
     }
 
-    private static void writeTemplateBetween(Writer writer, String template, String begin, String end) throws IOException {
+    private static void writeTemplateBetween(Writer writer, String template, String begin, String end)
+            throws IOException {
         int beginIndex = begin == null ? 0 : template.indexOf(begin) + begin.length();
         int endIndex = end == null ? template.length() : template.indexOf(end);
         writer.write(template.substring(beginIndex, endIndex));
@@ -120,4 +134,12 @@ public final class MessagesToHtmlWriter implements AutoCloseable {
         }
         return new String(baos.toByteArray(), UTF_8);
     }
+
+    @FunctionalInterface
+    public interface Serializer {
+
+        void writeValue(Writer writer, Envelope value) throws IOException;
+
+    }
+
 }